仮登録バックエンド処理　完成
