+++ application/config/development/routes.php
@@ -49,6 +49,6 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 | Examples:	my-controller/index	-> my_controller/index
 |		my-controller/my-method	-> my_controller/my_method
 */
-$route['default_controller'] = 'cl_main';
+$route['default_controller'] = 'cl_landing';
 $route['404_override'] = '';
 $route['translate_uri_dashes'] = FALSE;
