+++ application/views/forgot-password.php
