+++ application/views/index.html
@@ -88,7 +88,7 @@
                         <div class="item-content-inner">
                             <h2 class="heading"> <br class="d-none d-md-block">ﾍﾟｯﾄｻﾛﾝ専用CRM</h2>
                             <p class="intro">無料で始めるシンプルな顧客＆ﾍﾟｯﾄ管理</p>
-                            <a class="btn btn-primary btn-cta" href="cl_main/login" target="_blank">会員登録して始める</a>
+                            <a class="btn btn-primary btn-cta" href="../cl_landing/login" target="_blank">会員登録して始める</a>
                         </div>
                         <!--//item-content-inner-->
                     </div>
