+++ application/views/view_registration_mail.php
@@ -6,7 +6,7 @@
     <meta content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" name="viewport">
     <title>Animarl</title>
     <!-- Favicon-->
-    <link rel="icon" href="favicon.ico" type="image/x-icon">
+    <link rel="icon" href="../favicon.ico" type="image/x-icon">
 
     <!-- Google Fonts -->
     <link href="https://fonts.googleapis.com/css?family=Roboto:400,700&subset=latin,cyrillic-ext" rel="stylesheet"
@@ -26,48 +26,37 @@
     <link href="../assets/cms/css/style.css" rel="stylesheet">
 </head>
 
-<body class="login-page">
-    <div class="login-box">
+<body class="signup-page">
+    <div class="signup-box">
         <div class="logo">
             <a href="javascript:void(0);"><b>Animarl</b></a>
+            <small>新規会員登録</small>
+            <!-- <small>Admin BootStrap Based - Material Design</small> -->
         </div>
         <div class="card">
             <div class="body">
-                <form action="../cl_login/check_user" method="POST">
-                    <div class="msg">メールアドレスとパスワードを入力してください</div>
+                <form action="../cl_register/register_email" method="POST">
+                    <div class="msg">まず初めにEmailを登録してください</div>
+                    <?php echo form_error('email');
+                        isset($error)? print($error): false;
+                    ?>
                     <div class="input-group">
                         <span class="input-group-addon">
                             <i class="material-icons">email</i>
                         </span>
                         <div class="form-line">
-                            <input type="email" class="form-control" name="email" placeholder="Email" required
-                                autofocus>
+                            <input type="email" class="form-control" name="email" placeholder="メールアドレス" required>
                         </div>
                     </div>
-                    <div class="input-group">
-                        <span class="input-group-addon">
-                            <i class="material-icons">lock</i>
-                        </span>
-                        <div class="form-line">
-                            <input type="password" class="form-control" name="password" placeholder="Password" required>
-                        </div>
-                    </div>
-                    <div class="row">
-                        <div class="col-xs-8 p-t-5">
-                            <input type="checkbox" name="rememberme" id="rememberme" class="filled-in chk-col-pink">
-                            <label for="rememberme">Remember Me</label>
-                        </div>
-                        <div class="col-xs-4">
-                            <button class="btn btn-block bg-pink waves-effect" type="submit">SIGN IN</button>
-                        </div>
-                    </div>
-                    <div class="row m-t-15 m-b--20">
-                        <div class="col-xs-5">
-                            <a href="signup">新規会員登録</a>
-                        </div>
-                        <div class="col-xs-7 align-right">
-                            <a href="forgot_password">パスワードを忘れた</a>
-                        </div>
+                    <!-- <div class="form-group">
+                        <input type="checkbox" name="terms" id="terms" class="filled-in chk-col-pink">
+                        <label for="terms">I read and agree to the <a href="javascript:void(0);">terms of
+                                usage</a>.</label>
+                    </div> -->
+                    <button class="btn btn-block btn-lg bg-pink waves-effect" type="submit">SIGN UP</button>
+
+                    <div class="m-t-25 m-b--5 align-center">
+                        <a href="login">アカウントを持っていますか?</a>
                     </div>
                 </form>
             </div>
@@ -88,7 +77,7 @@
 
     <!-- Custom Js -->
     <script src="../assets/cms/js/admin.js"></script>
-    <script src="../assets/cms/js/pages/examples/sign-in.js"></script>
+    <script src="../assets/cms/js/pages/examples/sign-up.js"></script>
 </body>
 
 </html>
